backend/services/python/users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = "The users service"
authors = ["copper-boy <sermed512@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
pydantic = {extras = ["email"], version = "^1.10.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-jwt-auth = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
